use ekka 0.3.1
- with the new version of ekka, an emqttd node will keep trying to join the
cluster in a loop, and not just try one time when it boots.
- the new behavior is particularly interesting when emqttd is running in
a k8s environment, and nodes can scale up/down. In the eventuality that
two or more nodes start at the same time, and there is no existing nodes
in the cluster yet, there is a very high chance that one of nodes
doesn't join the cluster.